Added web stats on overview
